requirements.txt
numpy
Pillow
tensorflow==2.3.0
tensorflow-addons==0.11.1
tabulate
tqdm
packaging
onnxruntime
onnx==1.7.0
onnx-tf
netron
